amplify/functions/thumbnail-generator/requirements.txt
boto3>=1.34.0
Pillow>=10.0.0